+++ projects/journey-maps-client/src/lib/services/multiTouchSupport.ts
@@ -4,7 +4,7 @@
 export class MultiTouchSupport {
 
   private state: any;
-  private map: any;
+  private map: mapboxgl.Map;
   private container: any;
   private TOUCH_ZOOM_FACTOR = 0.2; // zoom sensibility
   private TOUCH_ZOOM_THRESHOLD = 0.01; // zoom activation